requirements.txt
Django==2.0.4
pandas==0.19.2
PyJWT==1.6.1
